requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.12.30
python-dotenv>=1.0.0
asyncio
aiohttp
PyNaCl>=1.4.0
youtube-dl
